Move contacts into feature based folder structure
This starts to address #74 by moving files into feature concerns
rather than type concerns.

It involves moving files and the references to them including starting
to create a basic router that is expected at the root of each sub-app.
