+++ config/routes/index.js
@@ -1,5 +1,7 @@
 const routers = []
 const controllers = [
+  '../../src/apps/contacts/router',
+
   '../../src/controllers/api.controller',
   '../../src/controllers/pingdom.controller',
   '../../src/controllers/login.controller',
@@ -7,10 +9,6 @@ const controllers = [
   '../../src/controllers/index.controller',
   '../../src/controllers/my-account.controller',
   '../../src/controllers/search.controller',
-  '../../src/controllers/contact.controller',
-  '../../src/controllers/contact-edit.controller',
-  '../../src/controllers/contact-interaction.controller',
-  '../../src/controllers/contact-archive.controller',
   '../../src/controllers/interaction.controller',
   '../../src/controllers/interaction-edit.controller',
   '../../src/controllers/service-delivery.controller',
