+++ src/apps/contacts/contact.repo.js
@@ -1,6 +1,6 @@
-const logger = require('../../config/logger')
-const authorisedRequest = require('../lib/authorised-request')
-const config = require('../../config')
+const logger = require('../../../config/logger')
+const authorisedRequest = require('../../lib/authorised-request')
+const config = require('../../../config')
 
 function getContact (token, contactId) {
   return authorisedRequest(token, `${config.apiRoot}/v3/contact/${contactId}`)
